Prepping for multiple battles
created enemy scenes and classes
tweaked passing in and getting battle song templates
