+++ Audio/335571__magntron__gamemusic_120bpm.mp3.import
@@ -13,7 +13,7 @@ dest_files=["res://.godot/imported/335571__magntron__gamemusic_120bpm.mp3-a87b35
 [params]
 
 loop=true
-loop_offset=0
-bpm=0
+loop_offset=0.0
+bpm=0.0
 beat_count=0
 bar_beats=4
