+++ scenes/Puppets/Enemies/BossBlood/assets/Boss1EmissionShape.tres
