+++ scenes/Puppets/Enemies/BossBlood/assets/Enemy1EmissionShape.png
