+++ scenes/Puppets/Enemies/BossBlood/assets/Enemy1EmissionShape.png.import
@@ -3,15 +3,15 @@
 importer="texture"
 type="CompressedTexture2D"
 uid="uid://b5jpi2i502lw5"
-path="res://.godot/imported/Enemy1EmissionShape.png-e1502c3313247b217f7d453af1069017.ctex"
+path="res://.godot/imported/Enemy1EmissionShape.png-b52ecd60b2d507b2c2133463b0dcbc08.ctex"
 metadata={
 "vram_texture": false
 }
 
 [deps]
 
-source_file="res://scenes/Puppets/Enemies/assets/Enemy1EmissionShape.png"
-dest_files=["res://.godot/imported/Enemy1EmissionShape.png-e1502c3313247b217f7d453af1069017.ctex"]
+source_file="res://scenes/Puppets/Enemies/BossBlood/assets/Enemy1EmissionShape.png"
+dest_files=["res://.godot/imported/Enemy1EmissionShape.png-b52ecd60b2d507b2c2133463b0dcbc08.ctex"]
 
 [params]
