+++ scenes/Puppets/Enemies/EnemyPuppet.cs
@@ -1,0 +1,7 @@
+using System;
+using Godot;
+
+public partial class EnemyPuppet : PuppetTemplate
+{
+    //TODO: What do enemies need?
+}
